refactor: Move new command implementation atop the now-emptied old location.
