+++ src/camera/white-balance.ts
@@ -1,5 +1,5 @@
 import { assertNever } from '@companion-module/base'
-import { ModuleDefinedCommand } from '../visca/newcommand.js'
+import { ModuleDefinedCommand } from '../visca/command.js'
 
 export type WhiteBalanceMode = 'automatic' | 'indoor' | 'outdoor' | 'onepush' | 'manual'
