+++ src/visca/__tests__/command-expands-to-include-terminator.test.ts
@@ -1,9 +1,9 @@
 import { InstanceStatus } from '@companion-module/base'
 import { describe, test } from '@jest/globals'
-import { UserDefinedCommand } from '../newcommand.js'
 import { CommandFailed, SendCommand, WaitUntilConnectedToCamera } from './camera-interactions/interactions.js'
 import { AttemptSendInvalidMatcher, MatchVISCABytes } from './camera-interactions/matchers.js'
 import { RunCameraInteractionTest } from './camera-interactions/run-test.js'
+import { UserDefinedCommand } from '../command.js'
 
 describe('terminator embedded in command via param', () => {
 	test('terminator in computed command bytes', async () => {
